Splitting skeleton notebook into 4
- Causal Discovery notebooks
- Identification and Estimation notebook
- Refute notebook
- Policy Recommender notebook
