+++ graph.txt
@@ -1,0 +1,1 @@
+digraph  { dpi=200; fontsize=18; "Global Flag" [label="Global Flag"]; "Global Flag" [label="Global Flag"]; "Major Flag" [label="Major Flag"]; "Major Flag" [label="Major Flag"]; "SMC Flag" [label="SMC Flag"]; "SMC Flag" [label="SMC Flag"]; "Commercial Flag" [label="Commercial Flag"]; "Commercial Flag" [label="Commercial Flag"]; "IT Spend" [label="IT Spend"]; "IT Spend" [label="IT Spend"]; "Employee Count" [label="Employee Count"]; "Employee Count" [label="Employee Count"]; "PC Count" [label="PC Count"]; "PC Count" [label="PC Count"]; Size [label=Size]; Size [label=Size]; "Tech Support" [label="Tech Support"]; "Tech Support" [label="Tech Support"]; Discount [label=Discount]; Discount [label=Discount]; "New Engagement Strategy" [label="New Engagement Strategy"]; "New Engagement Strategy" [label="New Engagement Strategy"]; "New Product Adoption" [label="New Product Adoption"]; "New Product Adoption" [label="New Product Adoption"]; "Planning Summit" [label="Planning Summit"]; "Planning Summit" [label="Planning Summit"]; Revenue [label=Revenue]; Revenue [label=Revenue]; "Global Flag" -> Revenue  [arrowhead=normal, arrowtail=none, dir=both]; "Major Flag" -> "Tech Support"  [arrowhead=normal, arrowtail=none, dir=both]; "Major Flag" -> Discount  [arrowhead=normal, arrowtail=none, dir=both]; "Major Flag" -> "New Engagement Strategy"  [arrowhead=normal, arrowtail=none, dir=both]; "Major Flag" -> Revenue  [arrowhead=normal, arrowtail=none, dir=both]; "SMC Flag" -> "Tech Support"  [arrowhead=normal, arrowtail=none, dir=both]; "SMC Flag" -> Discount  [arrowhead=normal, arrowtail=none, dir=both]; "SMC Flag" -> "New Engagement Strategy"  [arrowhead=normal, arrowtail=none, dir=both]; "Commercial Flag" -> "Tech Support"  [arrowhead=normal, arrowtail=none, dir=both]; "Commercial Flag" -> Discount  [arrowhead=normal, arrowtail=none, dir=both]; "Commercial Flag" -> "New Engagement Strategy"  [arrowhead=normal, arrowtail=none, dir=both]; "Commercial Flag" -> Revenue  [arrowhead=normal, arrowtail=none, dir=both]; Size -> "IT Spend"  [arrowhead=normal, arrowtail=none, dir=both]; "IT Spend" -> "Tech Support"  [arrowhead=normal, arrowtail=none, dir=both]; "IT Spend" -> Discount  [arrowhead=normal, arrowtail=none, dir=both]; "IT Spend" -> "New Engagement Strategy"  [arrowhead=normal, arrowtail=none, dir=both]; "IT Spend" -> Revenue  [arrowhead=normal, arrowtail=none, dir=both]; "Employee Count" -> "PC Count"  [arrowhead=normal, arrowtail=none, dir=both]; "Employee Count" -> "Tech Support"  [arrowhead=normal, arrowtail=none, dir=both]; "Employee Count" -> Discount  [arrowhead=normal, arrowtail=none, dir=both]; "Employee Count" -> "New Engagement Strategy"  [arrowhead=normal, arrowtail=none, dir=both]; "PC Count" -> "Tech Support"  [arrowhead=normal, arrowtail=none, dir=both]; "PC Count" -> Discount  [arrowhead=normal, arrowtail=none, dir=both]; "PC Count" -> "New Engagement Strategy"  [arrowhead=normal, arrowtail=none, dir=both]; Size -> Revenue  [arrowhead=normal, arrowtail=none, dir=both]; "Tech Support" -> "New Product Adoption"  [arrowhead=normal, arrowtail=none, dir=both]; "Tech Support" -> "Planning Summit"  [arrowhead=normal, arrowtail=none, dir=both]; "Tech Support" -> Revenue  [arrowhead=normal, arrowtail=none, dir=both]; Discount -> Revenue  [arrowhead=normal, arrowtail=none, dir=both]; "New Engagement Strategy" -> "Planning Summit"  [arrowhead=normal, arrowtail=none, dir=both]; "New Product Adoption" -> Revenue  [arrowhead=normal, arrowtail=none, dir=both]; Revenue -> "Planning Summit"  [arrowhead=normal, arrowtail=none, dir=both]; }
