+++ util/notebook-config.py
@@ -15,6 +15,10 @@ dbutils.library.restartPython()
 
 # COMMAND ----------
 
+# MAGIC %matplotlib inline
+
+# COMMAND ----------
+
 def add_directions(causal_graph, directions):
     for direction in directions:
         causal_graph.G.add_directed_edge(
@@ -44,23 +48,32 @@ def add_relations_influencing_incentives(
 
 # COMMAND ----------
 
+from pip._vendor import pkg_resources
+
+def get_version(package):
+    package = package.lower()
+    return next((p.version for p in pkg_resources.working_set if p.project_name.lower() == package), "No match")
+
+# COMMAND ----------
+
+# ignoring deprication warnings
+from numba.core.errors import NumbaDeprecationWarning, NumbaPendingDeprecationWarning
+import warnings
+
+warnings.simplefilter("ignore", category=NumbaDeprecationWarning)
+warnings.simplefilter("ignore", category=NumbaPendingDeprecationWarning)
+
 # Generic ML imports
 from sklearn.pipeline import make_pipeline, Pipeline
 from sklearn.compose import ColumnTransformer
 from sklearn.preprocessing import StandardScaler
 from sklearn.linear_model import LogisticRegression, Lasso
 
-# ignoring deprication warnings
-from numba.core.errors import NumbaDeprecationWarning, NumbaPendingDeprecationWarning
-import warnings
-
 # EconML imports
 from econml.dml import LinearDML
 from econml.cate_interpreter import SingleTreePolicyInterpreter
 
 def setup_treatment_and_out_models():
-  warnings.simplefilter("ignore", category=NumbaDeprecationWarning)
-  warnings.simplefilter("ignore", category=NumbaPendingDeprecationWarning)
 
   # transformer that performs standard scaling on non-binary variables
   ct = ColumnTransformer(
@@ -133,7 +146,84 @@ def to_pydot(G, labels, dpi = 200):
 
 # COMMAND ----------
 
-def compare_returns_from_policies(final_df):
+import econml
+import dowhy
+import mlflow
+import sklearn
+
+class ModelWrapper(mlflow.pyfunc.PythonModel):
+    def __init__(self, model, estimand, estimate):
+        self._model = model
+        self._estimand = estimand
+        self._estimate = estimate
+
+    def get_model(self):
+        return self._model
+
+    def get_estimand(self):
+        return self._estimand
+
+    def get_estimate(self):
+        return self._estimate
+
+    def predict(self, context, model_input):
+        return self._estimate._estimator_object.const_marginal_effect(model_input)
+
+
+def get_model_env():
+    model_env = mlflow.pyfunc.get_default_conda_env()
+    model_env["dependencies"][-1]["pip"] += [
+        f"dowhy=={dowhy.__version__}",
+        f"econml=={econml.__version__}",
+        f"sklearn=={sklearn.__version__}",
+    ]
+    return model_env
+
+
+def register_dowhy_model(model_name, model, estimand, estimate):
+    with mlflow.start_run(run_name=f"{model_name} run") as run:
+        model_info = mlflow.pyfunc.log_model(
+            artifact_path="model",
+            python_model=ModelWrapper(
+                model=model,
+                estimand=estimand,
+                estimate=estimate,
+            ),
+            conda_env=get_model_env(),
+        )
+
+    return mlflow.register_model(
+        model_uri=f"runs:/{run.info.run_id}/model", name=model_name
+    )
+
+# COMMAND ----------
+
+import mlflow
+
+def get_registered_wrapped_model(model_name):
+  client = mlflow.MlflowClient()
+  latest_model_versions = client.get_latest_versions(
+      name=model_name
+  )
+
+  if len(latest_model_versions) >0:
+    latest_model_version = latest_model_versions[0].version
+  else:
+    raise Exception(f"There are no registered versions for model {model_name}")  
+
+  wrapped_model = mlflow.pyfunc.load_model(
+      f"models:/{model_name}/{latest_model_version}"
+  )
+
+  return wrapped_model.unwrap_python_model()
+
+def get_registered_wrapped_model_estimator(model_name):
+  wrapped_model = get_registered_wrapped_model(model_name=model_name)
+  return wrapped_model.get_estimate()._estimator_object  
+
+# COMMAND ----------
+
+def compare_policies_effects(final_df):
   average_effect_recommended_incentive_df = pd.DataFrame(
       final_df[
           [
